Changes actor identification from workflow ownership to thread ownership
Workflow previously identifies the actor to block from the actor attribute
assigned when a workflow instance is allocated to an actor.  However, it is
better if actors are associated with threads rather than workflows, since
otherwise a thread could become blocked by accessing an attribute in a
workflow blocked by another actor.